requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.30.1
tensorflow==2.18.0
pillow==10.3.0
numpy==1.26.4
streamlit==1.39.0